+++ pelutils/misc/_conditional_import.py
@@ -19,13 +19,3 @@ def import_git():
         return git
     except ImportError:  # The Python git library should always be available, but git may not be available on the system
         return None
-
-
-def import_pandas():
-    """Attempt a import of pandas. If found, pandas is returned, otherwise None is returned."""
-    try:
-        import pandas  # noqa: PLC0415
-
-        return pandas
-    except ModuleNotFoundError:
-        return None